remove: AssetSlot 클래스를 SlotEntry로 교체 및 관련 코드 삭제
- `AssetSlot` 클래스 및 관련 소스/헤더 파일 삭제
- `SlotEntry`로 재구성된 API를 활용하여 관련 코드 업데이트
- 테스트 코드에서 `AssetSlot` 관련 테스트 제거 및 `SlotEntry` 기반 테스트 유지
- 더 이상 사용되지 않는 `ELoadingState`, `AssetSlot`, `DeserializeAssetPayload_DEPRECATED` 제거
- 주석 업데이트 및 관련 코드 정리로 가독성 향상
